requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3